Fix: local build script, og image upgrade and astro 3.0 request name change

+++ scripts/build.js
@@ -1,5 +1,9 @@
 const fs = require('fs');
 
-// delete first the posts folder from the public/assets folder, then move the posts folder from root to public/assets
-fs.rmSync('public/assets/posts', { recursive: true });
+const postsFolderExists = fs.existsSync('public/assets/posts');
+
+if (postsFolderExists) {
+  fs.rmSync('public/assets/posts', { recursive: true });
+}
+
 fs.renameSync('posts', 'public/assets/posts');
